[CL-2210] Move component utils
